Enforce supported versions of node and npm
This package uses node features which require node v6.x and higher.
Prior to this change the package would install but a user would
get runtime errors for constructs (e.g. class) which are not
natively supported in prior version of node.  For compatibility with
prior versions of npm < 3.0.0, included both the "engine-strict" and
"engineStrict" (deprecated) flags.

Fixes FAB-725

Change-Id: Ib7042ecdec51bdcfddbd6be23f69d5391519572d
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>