FAB-1417 Move peers from request.targets to Chain
With the new SDK design, the Chain owns the peers
so the code needs to access the peers from the
the Chain, rather than passing them on the request.
This affects both sendDeploymentProposal and
sendTransactionProposal and both use
_checkProposalRequest, so changes to both functions
are required to allow the headless-tests to
continue to run successfully.

patch set 2: Fix gulp errors and update README.md
with instructions to use Mac natively to build
Docker images.

Change-Id: I48fc30cfcd9c1c688eed2bc820eb8f2f88f6af52
Signed-off-by: cdaughtr <cdaughtr@us.ibm.com>