[FAB-678] Omit dockerfile in deployment payload
Our deployment payloads carry Dockerfile contents in most
contexts.  This is problematic for a variety of reasons.

For example, it assumes that the client is in a position to
understand what docker images are available to each and every
peer in the network.  Consider that some peers may have
different image availability or be running on different
architectures.

In addition, the client is then responsible for installing
a meaningful peer.crt without having any knowledge of the peers
the chaincode will need to execute against.

It is also a potential source of attack and/or misconfiguration
since it provides another method to execute code within the
endorser network outside of the normal constraints provided by
the chaincode execution environment.

This patch is a step towards fixing FAB-678 by removing
the Dockerfile from our deployment payload. The peer will
provide a Dockerfile on behalf of the chaincode in a JIT
fashion.

Change-Id: I0951b28e11f764020db035f846afb848e694ae29
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>