Allow fabric-ca-client to use per-instance crypto
FAB-2497
allow crypto parameters (keysize, hash algo, etc.) to be passed
into the constructor of the client instance, rather than requiring
the app to call setConfigSetting() which is clunky and would
impact global settings.

as part of this also need to allow a hash function to be passed
into the "sign()" method in the SigningIdentity class because
the same signing identity may need to use different hashing
algorithms (like when talking to fabric-ca vs. peer/orderers)

Change-Id: Id21fe7c2b5b120f11f04cdd31f624fce687e4361
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>