NodeSDK - return error object from proposals
Return the result of an endorsement to the caller
if the call fails. If the result is not already an
Error object , wrap the failed result in an Error
Object. Remove the result.error wrapping so that
the Error is the actual result and may be tested for
existence.
Update latest orderer/configuration proto

Change-Id: Ic478353cb7a146dbbf0038c074be7bcfd4c45dc2
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>