Add tutorials to generated API docs
FAB-2817
This enables the `gulp doc` command to generate contents with
the tutorials written in the .md files under the "docs" folder.
Doing this allows us to have a single consistently style and
navigation for all of the SDK's documentation contents, including
the hand-written ones (tutorials) and jsdoc ones.

To preview how this gets renders, see:

https://jimthematrix.github.io/

Change-Id: If54c147f4cb358ae5d4a11f14304d6832018a318
Signed-off-by: Jim Zhang <jzhang@us.ibm.com>