[FAB-3089] NodeSDK add create config from JSON
Add support for using a simplified JSON based
description of a Channel configuration. This is
the start of NodeSDK being able to create a
channel without the use of configtx tool. The
new code will build the main config update protobuf
object needed by the Orderer to creat a channel
and the Organizations to sign. The next step will
be to add API's for the user to complete the full
user flow of the create channel.

Change-Id: I72b6e6a1b842565ebd33b722151e33e0f3717a96
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>