[FAB-2843] NodeSDK - Handle network issues part1
The first phase of handling network issues will be
for the event hub to report to the application that
the event stream has been shutdown. The contract with
the applications now includes an error callback similar
to the Promise resolve and reject.
Block decoding was enhanced so that the decoding of the
event block could be shared by all three event types
we provide, block, transaction, and chaincode.
Enhanced the jsDoc of Block.js.

Change-Id: I72082693e38bf4635b5f139f43205aabee15909d
Signed-off-by: Bret Harrison <beharrison@nc.rr.com>