[FAB-8229] Set orderer's name from network profile
This patchset changes NetworkConfig_1_0.getOrderer() to set the key
of the orderer configurations as the name of Orderer it returns.
This behavior matches that of peers.

This patchset also adds two test cases for network config regarding
orderers, and reenables network config unit tests, which might have
been skipped accidentally.

Change-Id: I0306cb6a92fbf5778393d336cff97663b7dc8a8b
Signed-off-by: Taku Shimosawa <taku.shimosawa@hal.hitachi.com>