[FAB-7654] Run PKCS11 tests w/o root privileges
The README currently describes how to run the
PKCS11 tests locally with SoftHSM. By default,
SoftHSM requires root privileges in order to run
the required commands.  In order to allow users
without root privileges to run the tests, a
config file has been added and the README was
updated accordingly.

Change-Id: Ie47ba3202e62665b5aa8bcef73d593eb6d1b7d22
Signed-off-by: Gari Singh <gari.r.singh@gmail.com>